[frontend] refactor: datatable
allow datatable to be configurable (sortable, paginated);
abstract out datatable header
